requirements.txt
docker
Click
pyyaml